Window: getX and getY returns uint16_t instead of uint32_t

+++ src/main/gui/windows/window.cpp
@@ -15,12 +15,12 @@ const Coord& Window::getCoord() const
 	return coord;
 }
 
-uint32_t Window::getX() const
+uint16_t Window::getX() const
 {
 	return coord.x;
 }
 
-uint32_t Window::getY() const
+uint16_t Window::getY() const
 {
 	return coord.y;
 }
